docs(core/content-header): rename to match property name
