+++ packages/angular-test-app/src/app/app-routing.module.ts
@@ -62,7 +62,7 @@ import Chip from '../preview-examples/chip';
 import ContentExample from '../preview-examples/content';
 import ContentHeader from '../preview-examples/content-header';
 import ContentHeaderNoBack from '../preview-examples/content-header-no-back';
-import ContentHeaderTruncate from '../preview-examples/content-header-truncate';
+import ContentHeaderTextOverflow from '../preview-examples/content-header-text-overflow';
 import ContentHeaderWithSlot from '../preview-examples/content-header-with-slot';
 import CustomField from '../preview-examples/custom-field';
 import CustomFieldValidation from '../preview-examples/custom-field-validation';
@@ -549,8 +549,8 @@ const routes: Routes = [
         component: ContentHeader,
       },
       {
-        path: 'content-header-truncate',
-        component: ContentHeaderTruncate,
+        path: 'content-header-text-overflow',
+        component: ContentHeaderTextOverflow,
       },
       {
         path: 'content-header-with-slot',
