+++ packages/angular-test-app/src/preview-examples/content-header-text-overflow.ts
@@ -8,11 +8,10 @@
  */
 
 import { Component } from '@angular/core';
-import { IxButton, IxContentHeader } from '@siemens/ix-angular/standalone';
 
 @Component({
+  standalone: false,
   selector: 'app-example',
-  imports: [IxContentHeader, IxButton],
-  templateUrl: './content-header-truncate.html',
+  templateUrl: './content-header-text-overflow.html',
 })
-export default class ContentHeaderTruncate {}
+export default class ContentHeaderTextOverflow {}
