+++ packages/html-test-app/src/preview-examples/content-header-text-overflow.html
@@ -12,7 +12,7 @@ LICENSE file in the root directory of this source tree.
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Content header truncation example</title>
+    <title>Content header text overflow example</title>
   </head>
   <body>
     <ix-content-header
